refactor: organise workflow JS logic into folders
